Move tool configs into `configs/`, bump to 0.0.13-beta.5
Slim the repo root by relocating configs that are referenced by
explicit path (so the move can't break tool auto-discovery):

- vitest.config.ts, vitest.e2e.config.ts, vitest.setup.ts
- tsconfig.{app,node,test}.json
- typedoc.json, typedoc-conf.js

`contributors.json` moves to `src/data/` so it's reachable via the
`@` alias instead of `@/../`. The duplicate `.babelrc` and
`babel.config.js` were both unused (Vite build is byte-identical
without them) and are removed.

Internal path references in the moved configs were rewritten to
account for their new depth, package.json scripts gained
`--config configs/...` flags, and `tsconfig.json` + `eslint.config.js`
now point at the relocated sub-tsconfigs. Tool auto-discovered configs
(vite.config.ts, eslint.config.js, root tsconfig.json, .prettierrc,
tsdoc.json, .npmrc, .nvmrc, .git*, index.html, config.json) stay in
root so IDE integrations and tools keep working without --config flags.

Verified: pnpm type-check, pnpm build, and pnpm test (1305 passing)
all green after the move.
