+++ configs/tsconfig.app.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
+    "tsBuildInfoFile": "../node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "target": "ES2024",
     "strictNullChecks": true,
     "useDefineForClassFields": true,
@@ -22,7 +22,7 @@
     "noEmit": true,
     "experimentalDecorators": true,
     "jsx": "react-jsx",
-    "baseUrl": "./",
+    "baseUrl": "../",
     "paths": {
       "@/*": [
         "./src/*"
@@ -36,8 +36,8 @@
     "noUncheckedSideEffectImports": true,
     "forceConsistentCasingInFileNames": true,
     "typeRoots": [
-      "./node_modules/@types",
-      "./src/types/**/*"
+      "../node_modules/@types",
+      "../src/types/**/*"
     ],
     "types": [
       "node",
@@ -45,7 +45,7 @@
     ]
   },
   "include": [
-    "./src"
+    "../src"
   ],
   "exclude": [
     "**/__tests__/**",
