+++ configs/tsconfig.test.json
@@ -1,0 +1,12 @@
+{
+  "extends": "./tsconfig.app.json",
+  "include": [
+    "../src/**/*.test.ts",
+    "../src/**/*.test.tsx",
+    "../src/**/*.spec.ts",
+    "../src/**/*.spec.tsx",
+    "../src/**/__tests__/**/*",
+    "../src/types/**/*.d.ts"
+  ],
+  "exclude": []
+}
