+++ configs/vitest.e2e.config.ts
@@ -4,10 +4,11 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   resolve: {
     alias: {
-      "@e2e": path.resolve(__dirname, "e2e"),
+      "@e2e": path.resolve(__dirname, "..", "e2e"),
     },
   },
   test: {
+    root: path.resolve(__dirname, ".."),
     include: ["e2e/**/*.e2e.test.ts"],
     testTimeout: 60_000,
     hookTimeout: 60_000,
