+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "chem-pal",
   "private": true,
-  "version": "0.0.13-beta.4",
+  "version": "0.0.13-beta.5",
   "type": "module",
   "engines": {
     "npm": ">=10.9.2",
@@ -17,14 +17,14 @@
     "build:full": "pnpm type-check && pnpm build",
     "lint": "eslint .",
     "preview": "vite preview",
-    "test": "vitest run",
+    "test": "vitest run --config configs/vitest.config.ts",
     "test:all": "pnpm test && pnpm test:e2e",
-    "test:ui": "vitest --ui --coverage --open=true",
-    "test:ui:open": "vitest --ui --coverage --open=true",
-    "test:coverage": "vitest run --coverage --open=true",
-    "test:run": "vitest run",
-    "test:e2e": "vitest run --config vitest.e2e.config.ts",
-    "docs": "typedoc --options typedoc.json && cp themes/typedoc-theme-hierarchy/packages/plugin/dist/assets/hierarchy.css themes/typedoc-theme-hierarchy/packages/plugin/dist/assets/hierarchy-theme.js docs/assets/",
+    "test:ui": "vitest --ui --coverage --open=true --config configs/vitest.config.ts",
+    "test:ui:open": "vitest --ui --coverage --open=true --config configs/vitest.config.ts",
+    "test:coverage": "vitest run --coverage --open=true --config configs/vitest.config.ts",
+    "test:run": "vitest run --config configs/vitest.config.ts",
+    "test:e2e": "vitest run --config configs/vitest.e2e.config.ts",
+    "docs": "typedoc --options configs/typedoc.json && cp themes/typedoc-theme-hierarchy/packages/plugin/dist/assets/hierarchy.css themes/typedoc-theme-hierarchy/packages/plugin/dist/assets/hierarchy-theme.js docs/assets/",
     "docs:www": "pnpm run docs && npx http-server -p 8080 docs",
     "validate-docs": "tsc --noEmit",
     "generate-logo-files": "node tools/generate-logo-files.js",
