+++ src/components/__tests__/AboutModal.test.tsx
@@ -1,4 +1,4 @@
-import contributors from "@/../contributors.json";
+import contributors from "@/data/contributors.json";
 import { fireEvent, render, screen } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import AboutModal from "../AboutModal";
