+++ tsconfig.json
@@ -3,14 +3,14 @@
   "references": [
     {
       // tsconfig.app.json is for your application source code (React components, business logic, etc.)
-      "path": "./tsconfig.app.json"
+      "path": "./configs/tsconfig.app.json"
     },
     {
       // tsconfig.node.json is specifically for Node.js environment files (like vite.config.ts, build scripts, etc.)
-      "path": "./tsconfig.node.json"
+      "path": "./configs/tsconfig.node.json"
     },
     {
-      "path": "./tsconfig.test.json"
+      "path": "./configs/tsconfig.test.json"
     }
   ],
   "compilerOptions": {
