--- tsconfig.test.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig.app.json",
-  "include": [
-    "src/**/*.test.ts",
-    "src/**/*.test.tsx",
-    "src/**/*.spec.ts",
-    "src/**/*.spec.tsx",
-    "src/**/__tests__/**/*",
-    "src/types/**/*.d.ts"
-  ],
-  "exclude": []
-}
